refactor: move backend tests to backend directory
