+++ backend/__tests__/sanitizeUser.test.js
